Refactor to move function to utils
