--- urbackupserver/www2/src/features/status/formatDatetime.tsx
@@ -1,14 +1,0 @@
-export function formatDatetime(datetime: number) {
-  const date = new Date(datetime * 1000);
-
-  const formatter = new Intl.DateTimeFormat("en-US", {
-    year: "numeric",
-    month: "2-digit",
-    day: "2-digit",
-    hour: "2-digit",
-    minute: "2-digit",
-    hour12: false,
-  });
-
-  return formatter.format(date);
-}
